requirements.txt
streamlit==1.31.1
psycopg2-binary==2.9.9
pandas==2.2.1
supabase==1.0.3